# lib/api-zod/src/generated/types/platformCount.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Samuga AI — centralized data ingestion, orchestration, and admin API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformCount {
  platform: string;
  count: number;
}