# lib/api-zod/src/generated/types/systemConfig.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Samuga AI — centralized data ingestion, orchestration, and admin API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemConfig {
  id: number;
  key: string;
  value: string;
  /** @nullable */
  description: string | null;
  /**
     * URL to push config changes to the external bot
     * @nullable
     */
  webhook_url: string | null;
  updatedAt: string;
}