# lib/api-zod/src/generated/types/telemetryEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Samuga AI — centralized data ingestion, orchestration, and admin API
 * OpenAPI spec version: 0.1.0
 */
import type { TelemetryEntryDataPayload } from './telemetryEntryDataPayload';

export interface TelemetryEntry {
  /** Name of the platform/app sending the event */
  platform: string;
  user_identifier: string;
  action_performed: string;
  /** Arbitrary JSON metadata */
  data_payload?: TelemetryEntryDataPayload;
}